pyproject.toml
[project]
name = "api-resnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "nest-asyncio>=1.6.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "pyngrok>=7.4.1",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
]